refactor: 상수파일 세분화
- error, domain, view 영역으로 나누어 상수파일 세분화 진행
