+++ src/constants/domain.js
@@ -1,0 +1,5 @@
+export const CAR = Object.freeze({
+  NAME_LENGTH_MIN: 1,
+  NAME_LENGTH_MAX: 5,
+  PROGRESS_CRITERIA: 4,
+});
