+++ src/constants/view.js
@@ -1,0 +1,9 @@
+export const INPUT_MESSAGE = Object.freeze({
+  CAR_NAMES: "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).",
+  TRY_COUNT: "시도할 횟수는 몇 회인가요?",
+});
+
+export const OUTPUT_MESSAGE = Object.freeze({
+  RESULT: "실행 결과",
+  PROGRESS_SYMBOL: "-",
+});
