+++ src/utils/createError.js
@@ -1,4 +1,4 @@
-import { ERROR_PREFIX } from "../constants/constants.js";
+import { ERROR_PREFIX } from "../constants/error.js";
 
 export const createError = (message) => {
   throw new Error(`${ERROR_PREFIX} ${message}\n`);
